fix: do not panic on binrw backtrace error
The ntfs crate could panic on some data because of a false
"unreachable" instruction in error.rs: the binrw error is
not guaranteed to only be an IO error, it can also be for
example a backtrace error.

Remove this unreachable which is too dangerous: any update of
binrw could make this unreachable wrong, it is better to be
conservative here.